app/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
Pillow
